First cut at a refactoring of resources so they're optional.  Compile with jam -NO=Resources to remove them.
git-svn-id: https://europa-pso.googlecode.com/svn/PLASMA/trunk@4244 23e1bc29-d954-0410-9766-47de34158ba0

+++ Utils/base/Utils.cc
@@ -5,7 +5,7 @@ DEFINE_GLOBAL_CONST(bool, g_alwaysFails, false);
 
 namespace EUROPA {
 
-  std::string toString(float value) {
+  std::string toString(double value) {
     std::stringstream s;
     s << value;
     return(s.str());
